Do not sort properties.
It is quite bad when typedoc reorders type struct properties.
This is most glaring for tagged unions such as `type S = {type: 'A', a: number} | {type: 'B', b: number}`
and for other structs where the definition order matters to people to understand the type better (e.g. optional fields at the end)